Changes resource source config keys.
Changed username -> client_name, password -> client_secret, path -> name.

This was done to match the namings of arguments that would be passed to
credhub cli for an equivalent operation.

(There are two different types of credentials that can be used to
authenticate with credhub- username and password (from uaa user)
and client and secret (from uaa client)- this resource
requires client and secret.